Faster inference with ngram probabilities stored in Vec
HashMap table lookups over the inner HashMap of langs and probs was
taking 60% of the inference time. Replacing this map with a vec gets rid
of these lookups. Each ngram map of probabilities was only being used to
check all its values, so it is not necessary to have a map. Iterate over
all the values of a vec is enough and faster.