Better handling of CJK
Only considered CJK chars the ones in the CJK_* blocks, which are only
Chinese and do not include Kana or Hangul. Not sure if it was the
intention, but it is the way HeLI does it.

Fixed a bug that was repeating the first non CJK character after CJK
group of characters.

More accurately compute the pct of cjk characters